Add syntax highlighting
Using Rainbow.js to highlight the <code> tag appended to the page. The
extension inserts Rainbow.js and its CSS into the blank document's head,
and manually appends a call to `Rainbow.color()`.

One current issue is that Rainbow.js actually appears to be _quite_ slow,
resulting in the extension being terminated for memory pressure on
large documents. This is against my expectations, as tossing a bunch of JS
at the page seemed like a great way to force Safari to deal with the
memory instead of the extension.

TODO: investigate a dozen other highlighting libraries